* fixes20200224: (30 commits)
  Test closing stdout early also through qrexec service
  Fix not closed file descriptors in qubes-rpc-multiplexer
  policy daemon: fix double ++ in logs
  gitignore: ignore 'build' dir
  Add '[Install]' section to qubes-qrexec-policy-daemon.service
  libqrexec: fix cleanup on qubes_connect error
  qrexec-client{,-vm}: do not fail if cannot sent just final EOF
  Use different error messages for different error paths
  Test stdio socket for client-vm as well
  Handle SIGPIPE in qrexec-client-vm
  Fix test_exit_before_closing_streams
  Remove debug print
  Add a test for replace_chars
  Split handle_new_process_common
  qrexec-client-vm: pass local process PID to process_io
  Always return remote exit status
  Fix timeout in wait_for_vchan_or_argfd
  Move all closing code to process_io.c; don't exit early
  Add a test in which we wait for a granchild to close streams
  Fix reading from a dead process (ECONNRESET on read)
  ...

+++ libqrexec/exec.c
@@ -187,10 +187,12 @@ static int qubes_connect(int s, const char *connect_path, const size_t total_pat
        result = connect(s, (struct sockaddr *) &remote, socket_len);
     while (result < 0 && (errno == EINTR || errno == EWOULDBLOCK || errno == EAGAIN));
     dummy_errno = errno;
-    unlink(buf);
 out:
-    buf[path_separator_offset] = '\0';
-    rmdir(buf);
+    if (buf[path_separator_offset] == '/') {
+        unlink(buf);
+        buf[path_separator_offset] = '\0';
+        rmdir(buf);
+    }
     errno = dummy_errno;
     return result;
 }
@@ -366,7 +368,7 @@ static int execute_parsed_qubes_rpc_command(
         *stdout_fd = *stdin_fd = s;
         if (stderr_fd)
             *stderr_fd = -1;
-        *pid = -1;
+        *pid = 0;
         set_nonblock(s);
         buffer_append(stdin_buffer, command->service_descriptor, strlen(command->service_descriptor) + 1);
         return 0;
